refactor(app): split App.tsx into modular components and hooks
- Extract usePhotoState, useModalState, usePhotoNavigation hooks
- Create Toolbar, Filmstrip, StatusBar, EmptyState components
- Reduce App.tsx from 950+ to ~400 lines
